GMINS-3.0.3/gmins_gui/test_port_detection.py
#!/usr/bin/env python3
"""
Test script for COM port detection functionality
Tests the port detection without GUI dependencies
"""

import serial.tools.list_ports
import sys
import os


def test_basic_port_detection():
    """Test basic serial port detection"""
    print("🔍 Testing basic COM port detection...")
    print("=" * 50)
    
    try:
        ports = serial.tools.list_ports.comports()
        print(f"System detected {len(ports)} COM ports:")
        
        if not ports:
            print("⚠️  No COM ports detected")
            print("   This could mean:")
            print("   - No USB serial devices connected")
            print("   - Device drivers not installed")
            print("   - Ports in use by other applications")
        else:
            for i, port in enumerate(ports, 1):
                print(f"  {i}. {port.device}")
                print(f"     Description: {port.description}")
                if port.manufacturer:
                    print(f"     Manufacturer: {port.manufacturer}")
                if port.vid and port.pid:
                    print(f"     VID:PID = {port.vid:04X}:{port.pid:04X}")
                print()
        
        return ports
        
    except Exception as e:
        print(f"❌ Error during port detection: {e}")
        return []


def test_port_info_structure():
    """Test the port information structure matches our GUI expectations"""
    print("📋 Testing port info structure...")
    print("=" * 50)
    
    ports = serial.tools.list_ports.comports()
    
    if not ports:
        # Create mock port for testing structure
        print("Creating mock port data for structure test...")
        mock_ports = [{
            'device': '/dev/ttyUSB0',
            'description': 'USB Serial Device',
            'manufacturer': 'FTDI',
            'vid': 0x0403,
            'pid': 0x6001
        }]
        
        print("Mock port structure:")
        for port in mock_ports:
            print(f"  Device: {port['device']}")
            print(f"  Description: {port['description']}")
            print(f"  Manufacturer: {port['manufacturer']}")
            print(f"  Display format: {port['device']} - {port['description']}")
            print()
    else:
        print("Real port structure:")
        for port in ports:
            port_dict = {
                'device': port.device,
                'description': port.description,
                'manufacturer': port.manufacturer or 'Unknown',
                'vid': port.vid,
                'pid': port.pid
            }
            print(f"  Device: {port_dict['device']}")
            print(f"  Description: {port_dict['description']}")
            print(f"  Manufacturer: {port_dict['manufacturer']}")
            print(f"  Display format: {port_dict['device']} - {port_dict['description']}")
            print()


def simulate_gui_behavior():
    """Simulate the GUI combo box behavior"""
    print("🖥️  Simulating GUI combo box behavior...")
    print("=" * 50)
    
    ports = serial.tools.list_ports.comports()
    
    # Simulate combo box items
    combo_items = []
    
    if not ports:
        print("No ports detected - would show:")
        combo_items.append(("No COM ports detected", ""))
        print("  'No COM ports detected' (data: '')")
    else:
        print("Would populate combo box with:")
        for port in ports:
            display_text = f"{port.device} - {port.description}"
            device_name = port.device
            combo_items.append((display_text, device_name))
            print(f"  '{display_text}' (data: '{device_name}')")
    
    print(f"\nTotal combo box items: {len(combo_items)}")
    
    # Simulate connection attempt
    if combo_items and combo_items[0][1]:  # If we have a valid port
        selected_port = combo_items[0][1]
        selected_display = combo_items[0][0]
        print(f"\nSimulated connection attempt:")
        print(f"  Selected port: {selected_port}")
        print(f"  Display text: {selected_display}")
        print(f"  Would attempt connection to: {selected_port}")
    else:
        print(f"\nNo valid ports for connection")


def main():
    print("🚀 GMINS GUI COM Port Detection Test")
    print("="*60)
    print()
    
    # Run tests
    test_basic_port_detection()
    print()
    test_port_info_structure()
    print()
    simulate_gui_behavior()
    print()
    
    print("✅ Port detection test completed!")
    print("\n💡 Next steps:")
    print("   1. Connect a USB serial device to see real ports")
    print("   2. Run the GUI to test the integrated functionality")
    print("   3. Use 'Refresh Ports' menu option to update the list")


if __name__ == "__main__":
    main()